Get information about a binary dosage file

Routine to return information about a binary dosage file. This information is used by other routines to allow for quicker extraction of values from the file.

getbdinfo(bdfiles)

Arguments

bdfiles
Vector of file names. The first is the binary dosage data containing the dosages and genetic probabilites. The second file name is the family information file. The thrid file name is the SNP information file. The family and SNP information files are not used if the binary dosage file is in format 4. For this format the family and SNP information are in the file with the dosages and genetic probabilites.

Value

List with information about the binary dosage file. This includes family and subject IDs along with a list of the SNPs in the file. Other information needed to read the file is also included.

Examples

vcf1abdfile <- system.file("extdata", "vcf1a.bdose", package = "BinaryDosage")
bdinfo <- getbdinfo(bdfiles = vcf1abdfile)